requirements.txt
selenium==4.8.3
notifiers==1.3.3
fake-useragent==1.1.3
psycopg2-binary==2.9.6
